+++ src/main/java/roomescape/dto/response/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.response;
 
 import java.time.LocalTime;
 import roomescape.domain.ReservationTime;
