+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -14,8 +14,8 @@ import org.junit.jupiter.api.Test;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.dto.ReservationCreationRequest;
-import roomescape.dto.ReservationResponse;
+import roomescape.dto.request.ReservationCreationRequest;
+import roomescape.dto.response.ReservationResponse;
 import roomescape.exception.BadRequestException;
 import roomescape.exception.NotFoundException;
 import roomescape.repository.ReservationRepository;
